Update autocompleter.cpp
I just added some stuff that you were missing and I added two functions in the class because there are four cases for rotating. That is what I got for insert. I just left completions recurse the same. I think we have the almost the same stuff for that part and yea.